test: integrate array coverage into GraphQL builder tests
